RichMenu関係を別クラスに移動
